% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_proximity.R
\name{avg_proximity}
\alias{avg_proximity}
\title{avg_proximity}
\usage{
avg_proximity(from_poly, to_points, tolerance, group, pop_weights = NULL)
}
\arguments{
\item{from_poly}{A spatial polygon layer}

\item{to_points}{A spatial points layer that represents the
environmental hazards.}

\item{tolerance}{The maximum search distance for environmental hazards.}

\item{group}{Name of variable in the from_poly layer that defines the group means
are calculated for.}

\item{pop_weights}{An optional vector of population weights if
weighted-averages are desired.}
}
\description{
Function to calculate average proximity to point-source
environmental hazards within a specified area. The function
will calculate a population weighted average if the user
supplies a vector of population counts or weights.
}
