% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ga}
\alias{ga}
\title{Georgia Census Tracts}
\format{
\subsection{\code{gabgrps}}{

A spatial data frame with 2,791 rows and 5 columns:
\describe{
\item{STATE}{State abbreviation}
\item{GEOID}{11-digit unique identifier of the tract}
\item{LSAD}{Year}
\item{POP}{Estimate of total population}
\item{geometry}{geometry}
...
}
}
}
\source{
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/cartographic-boundary.html}

\url{https://www.census.gov/programs-surveys/acs/data.html}
}
\usage{
ga
}
\description{
A spatial polygon dataset from the U.S. Census TIGER cartographic boundary files
(v2023). This dataset includes all 2020 census tracts in the state of Georgia. Data
are projected in NAD 1983 Georgia Statewide Lambert.
}
\details{
The dataset also includes a field containing total population counts from
2023 American Community Survey (ACS) 5-Year Estimates.
}
\seealso{
The script used to create the ga dataset: \url{https://github.com/cdcent/hazprox/blob/main/data-raw/}
}
\keyword{datasets}
