% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gabgrps}
\alias{gabgrps}
\title{Georgia Block Groups}
\format{
\subsection{\code{gabrgps}}{

A data frame with 7,441 rows and 5 columns:
\describe{
\item{STATE}{State abbreviation}
\item{GEOID}{12 digit unique identifier of the block group}
\item{LSAD}{Year}
\item{POP}{Estimate of total population}
\item{geometry}{geometry}
...
}
}
}
\source{
https://www.census.gov/geographies/mapping-files/time-series/geo/cartographic-boundary.html

https://www.census.gov/programs-surveys/acs/data.html
}
\usage{
gabgrps
}
\description{
A dataset from the U.S. Census TIGER cartographic boundary files (v2023) which
includes all 2020 block groups in the state of Georgia. Also includes
population counts from 2023 American Community Survey 5-Year Estimates.
}
\keyword{datasets}
