% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proximity.R
\name{get_proximity}
\alias{get_proximity}
\title{get_proximity}
\usage{
get_proximity(from_poly, to_points, tolerance)
}
\arguments{
\item{from_poly}{A spatial polygon layer}

\item{to_points}{A spatial points layer that represents the
environmental hazards.}

\item{tolerance}{The maximum search distance for environmental hazards.}
}
\description{
Function to calculate overall proximity to point-source
environmental hazards for areas. The function calculates
the sum of inverse distances between the geometric center
of the area and all hazards within a specified tolerance.
Proximity calculations for areas that do not have any hazards
within the specified tolerance only consider the single
nearest distance.
}
