% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazprox-package.R
\docType{package}
\name{hazprox-package}
\alias{hazprox}
\alias{hazprox-package}
\title{hazprox: Calculate Cumulative Proximity to Hazards for Geographic Areas}
\description{
A package for measuring cumulative proximity to hazards for geographic areas.
To learn more about hazprox, start with the vignettes:

\code{browseVignettes(package = "hazprox")}
}
\author{
\strong{Maintainer}: Sarah Rockhill \email{srockhill@cdc.gov} (\href{https://orcid.org/0000-0003-3197-2825}{ORCID})

}
\keyword{internal}
