% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_spatial.R
\name{is_spatial}
\alias{is_spatial}
\title{Function to verify that object is spatial data.}
\usage{
is_spatial(x)
}
\arguments{
\item{x}{}
}
\value{
Returns True if object is spatial and False otherwise.
}
\description{
Function to verify that object is spatial data.
}
