% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{npls}
\alias{npls}
\title{National Priority List/Superfund Sites}
\format{
\subsection{\code{npls}}{

A data frame with 17 rows and 13 columns:
\describe{
\item{NAME}{Name of Superfund site}
\item{EPA_ID}{A unique 12 digit identification number assigned by the Superfund Enterprise Management System}
\item{SITE_ID}{Unique identifier for site for Superfunds with multiple sites}
\item{STREET}{Street address of Superfund location}
\item{ADDR_C}{Additional address information for Superfund location, if any}
\item{CITY}{City/town/village/municipality of Superfund location or nearest place name if site is not within a formal jurisdiction.}
\item{COUNTY}{County of Superfund location}
\item{STATE}{State of Superfund location}
\item{ZIP}{5-digit zip code of Superfund location}
\item{DATE}{Date that information was queried from the EPA Envirofacts API}
\item{geometry}{geometry}
...
}
}
}
\source{
https://enviro.epa.gov/
}
\usage{
npls
}
\description{
A spatial points data frame containing National Priority List (otherwise known
as Superfund) sites in the state of Georgia or within 5 km of the state
boundary. Data are projected in NAD83(CSRS) / UTM zone 19N. Data were queried
from the EPA Envirofacts API on January 9, 2025.
}
\seealso{
The script used to create the npls dataset: \link{https://github.com/cdcent/hazprox/blob/main/data-raw/npls.R}
}
\keyword{datasets}
