% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_proximity.R
\name{avg_proximity}
\alias{avg_proximity}
\title{avg_proximity}
\usage{
avg_proximity(from, to, ..., group, pop_weights = NULL)
}
\arguments{
\item{from}{A spatial polygon layer with class sf or sfc. Proximity is calculated from
the geometric center of each polygon to each hazard within the search tolerance.}

\item{to}{A spatial polygon, point, or linestring layer with class sf or sfc
representing the environmental hazard(s) of interest.}

\item{...}{Additional parameters}

\item{group}{Name of grouping variable in the \code{from} layer. Average proximity is calculated
among polygons nested in the group.}

\item{pop_weights}{An optional vector of population weights if weighted-averages are desired.}
}
\description{
Function to calculate average proximity to environmental hazards within a specified area.
The function will calculate a population weighted average if the user supplies a vector
of population counts or weights.
}
\examples{
#find average census tract proximity to Superfund sites by county
ga$County <- substr(ga$GEOID, 1, 5)
counties <- avg_proximity(ga, npls,  group='County')
plot(counties['avg_prox'])
#find population-weighted average tract proximity to Superfund sites by county
counties_wt <- avg_proximity(ga, npls,  group='County', pop_weights = ga$POP)

}
