% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gabgrps}
\alias{gabgrps}
\title{Georgia Block Groups}
\format{
\subsection{\code{gabgrps}}{

A spatial data frame with 7,441 rows and 5 columns:
\describe{
\item{STATE}{State abbreviation}
\item{GEOID}{12 digit unique identifier of the block group}
\item{LSAD}{Year}
\item{POP}{Estimate of total population}
\item{geometry}{geometry}
...
}
}
}
\source{
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/cartographic-boundary.html}

\url{https://www.census.gov/programs-surveys/acs/data.html}
}
\usage{
gabgrps
}
\description{
A spatial polygon dataset from the U.S. Census TIGER cartographic boundary files
(v2023). This dataset includes all 2020 block groups in the state of Georgia. Data
are projected in NAD83(CSRS) / UTM zone 19N.
}
\details{
The dataset also includes a field containing total population counts from
2023 American Community Survey (ACS) 5-Year Estimates.
}
\seealso{
The script used to create the gabgrps dataset: \url{https://github.com/cdcent/hazprox/blob/main/data-raw/gabgrps.R}
}
\keyword{datasets}
