% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proximity.R
\name{get_proximity}
\alias{get_proximity}
\title{get_proximity}
\usage{
get_proximity(
  from_poly,
  to_points,
  tolerance = NULL,
  units = "km",
  weights = NULL
)
}
\arguments{
\item{from_poly}{An spatial polygon layer with class sf or sfc. Proximity
statistics will be calculated for each polygon in this layer.}

\item{to_points}{A spatial points layer with class sf or sfc
representing the environmental hazard of interest.}

\item{tolerance}{The maximum search distance for environmental hazards (optional).}

\item{units}{The units of the tolerance value (e.g., "m" ,  "km", "ft", "yd",  "fathom", "mi", "naut_mi", "au"),
if tolerance is specified.}

\item{weights}{An optional vector with the same length as to_points that applies a weight to each hazard.}
}
\description{
Function to calculate overall proximity to point-source
environmental hazards for areas. The function calculates
the sum of inverse distances between the geometric center
of the area and all hazards within a specified tolerance.
Proximity calculations for areas that do not have any hazards
within the specified tolerance only consider the single
nearest distance.
}
