% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proximity.R
\name{get_proximity}
\alias{get_proximity}
\title{Calculate cumulative proximity to features for geographic areas}
\usage{
get_proximity(from, to, tolerance = NULL, units = "km", weights = NULL)
}
\arguments{
\item{from}{polygon or multipolygon object with class sf or sfc.}

\item{to}{object of any geometry type (point, polygon, linestring, geometry collection) with class sf or sfc.}

\item{tolerance}{The maximum search distance for \code{to} features (optional).}

\item{units}{The units of the tolerance value (e.g., "m", "km", "ft", "yd", "fathom", "mi", "naut_mi", "au"),
if tolerance is specified. Ignored if tolerance is omitted.}

\item{weights}{An optional numeric vector with the same length as \code{to} that applies a weight to each \code{to} feature.}
}
\value{
A numeric vector with the same length as \code{from}.
}
\description{
This function calculates the sum of inverse distances between the geometric center
of the \code{from} area and all \code{to} features within a specified tolerance.
}
\details{
get_proximity calculates proximity statistics for each polygon in  \code{from} based on the cumulative inverse distance
to each feature in \code{to}, typically representing environmental hazard(s) of interest or concern. If a tolerance value
is provided, only hazards within the tolerance distance will be included in proximity calculations. For polygons that
do not have any hazards within the specified tolerance, proximity will only consider the single nearest distance.

Users may also weight the proximity calculations based on the expected risk or severity of each hazard in \code{to} by
providing a vector of weights corresponding to each hazard.
}
\examples{
set.seed(123)
w <- floor(runif(nrow(npls), min=0, max=10))
#calculate proximity to all superfund sites
p <- get_proximity(from=ga, to=npls, weights = w)
ga$Proximity <- p
plot(ga['Proximity'])
#calculate proximity to superfund sites within 10 miles of tract boundary
p10 <- get_proximity(from=ga, to=npls, tolerance = 10, units = 'mi', weights = w)

}
