% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_projection.R
\name{sync_projection}
\alias{sync_projection}
\title{Ensures that spatial data are in same projected coordinate system.}
\usage{
sync_projection(df1, df2)
}
\arguments{
\item{df1}{A spatial data frame}

\item{df2}{A spatial data frame}
}
\value{
A list object containing both spatial data frames projected in the same CRS.
}
\description{
Ensures that spatial data are in same projected coordinate system.
}
